requirements.txt
scoop
deap
scikit-learn
sklearn-genetic
statsmodels
matplotlib
numpy
scipy
pandas
ipython
networkx
